requirements.txt
Django>=3.0.4
djangorestframework>=3.11.0
psycopg2>=2.8.4
flake8>=3.7.4
requests>=2.23.0
uwsgi>=2.0.17
Pillow>=7.0.0